God, I've never seen such aids code.
