+++ src/main/java/me/leandrit0/ktk/events/events/StopEvent.java
@@ -1,0 +1,13 @@
+package me.leandrit0.ktk.events.events;
+
+import org.bukkit.entity.Player;
+
+import me.leandrit0.ktk.events.KillTheKingEvent;
+
+public class StopEvent extends KillTheKingEvent {
+     
+	public StopEvent(Player player) {
+		super(player);				
+	}
+
+}
