+++ src/main/java/me/leandrit0/ktk/utils/EncoderDecoder.java
