+++ src/main/java/me/leandrit0/ktk/utils/HiddenStringUtil.java
