+++ src/main/java/me/leandrit0/ktk/utils/JavaUtils.java
