--- src/me/leandrit0/ktk/events/type/Stop.java
@@ -1,13 +1,0 @@
-package me.leandrit0.ktk.events.type;
-
-import org.bukkit.entity.Player;
-
-import me.leandrit0.ktk.events.KillTheKingEvent;
-
-public class Stop extends KillTheKingEvent {
-     
-	public Stop(Player player) {
-		super(player);				
-	}
-
-}
