--- src/me/leandrit0/ktk/manager/KingManager.java
@@ -1,31 +1,0 @@
-package me.leandrit0.ktk.manager;
-
-import java.lang.reflect.InvocationTargetException;
-
-import org.bukkit.entity.Player;
-
-import lombok.Getter;
-import lombok.Setter;
-
-public class KingManager {
-	
-	@Getter @Setter public static Player king;
-	
-	public static boolean isActive() {
-		return king != null;
-	}
-	
-    public double getKingHealth() throws IllegalAccessException, IllegalArgumentException, InvocationTargetException, NoSuchMethodException, SecurityException {
-    	return  (double) king.getClass().getMethod("getHealth", double.class).invoke(king);
-    }
-    
-    public void removeKing(boolean isCancelled) {
-    	if (isCancelled) {
-    		king = null;
-    	}
-    }
-	
-
-	
-
-}
